requirements.txt
python-telegram-bot==20.6
pytz==2023.3
python-dotenv==1.0.0 